Need to inline solr params that reference '$q'
While items in the config under `search_attr_defaults` are being munged
so any references to $q, $qq, or $t are changed to the appropriate numbered
version for that edismax clause, the same is _not_ true of the generic
stuff under `solr_param`. This means that an attempt to take the definition
of the allfields boost outside of the allfields config (and then just refering
to it as $boost1) failed -- the $q in the boost query wasn't being changed and
ended up with the value `_query_:{!edismax ... }`.